NPC dialogue work
The NPC text box is better and it moves now but something is still wrong, it
scrolls in a weird way but the action levels seem okay. I think the appraoch of
correcting the textbox location based on cam is the best way to do it, I'm just
missing something.